requirements.txt
lightning
pytorch_lightning
torchmetrics[image]==0.11.4
torchgeometry

numpy==1.23.0
scipy
pandas
scikit-learn
scikit-image
joblib>=1.2.0
einops

matplotlib
tqdm
colorama
termcolor
click
easydict
hydra-core~=1.3.2
omegaconf~=2.3.0

gym
pyvirtualdisplay
opencv-python==4.6.0.66
vtk
open3d
moviepy
imageio
imageio-ffmpeg
pyzmq
pyrealsense2

autograd
rl-games==1.1.4
diffusers
gluonts[torch]==0.13.1
pytorchvideo~=0.1.5
rotary_embedding_torch

pyyaml
gdown
internetarchive
h5py
chardet
cchardet
lxml
ipdb
wandb
wandb-osh==1.2.1

chumpy
human_body_prior
numpy-stl

git+https://github.com/openai/CLIP.git
git+https://github.com/ZhengyiLuo/smplx.git@master
git+https://github.com/ZhengyiLuo/SMPLSim.git@master
